move code out of inner-classes into first-class classes
