+++ geode-tcp-server/src/main/java/org/apache/geode/distributed/internal/tcpserver/TcpSocketCreatorImpl.java
@@ -15,15 +15,8 @@
 package org.apache.geode.distributed.internal.tcpserver;
 
 import java.io.IOException;
-import java.net.BindException;
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
-import java.net.ServerSocket;
 import java.net.Socket;
-import java.nio.channels.ServerSocketChannel;
-import java.util.concurrent.ThreadLocalRandom;
 
-import org.apache.geode.util.internal.GeodeGlossary;
 
 /**
  * TcpSocketCreatorImpl is a simple implementation of TcpSocketCreator for use in the
@@ -33,32 +26,18 @@ import org.apache.geode.util.internal.GeodeGlossary;
 public class TcpSocketCreatorImpl implements TcpSocketCreator {
 
 
-  public static final boolean ENABLE_TCP_KEEP_ALIVE;
-
-  static {
-    // bug #49484 - customers want tcp/ip keep-alive turned on by default
-    // to avoid dropped connections. It can be turned off by setting this
-    // property to false
-    String str = System.getProperty(GeodeGlossary.GEMFIRE_PREFIX + "setTcpKeepAlive");
-    if (str != null) {
-      ENABLE_TCP_KEEP_ALIVE = Boolean.valueOf(str);
-    } else {
-      ENABLE_TCP_KEEP_ALIVE = true;
-    }
-  }
-
-  protected ServerSocketCreator serverSocketCreator;
-  protected ClientSocketCreator clientSocketCreator;
-  protected AdvancedSocketCreator advancedSocketCreator;
+  protected ServerSocketCreatorImpl serverSocketCreator;
+  protected ClientSocketCreatorImpl clientSocketCreator;
+  protected AdvancedSocketCreatorImpl advancedSocketCreator;
 
   public TcpSocketCreatorImpl() {
     initializeCreators();
   }
 
   protected void initializeCreators() {
-    serverSocketCreator = new ServerSocketCreatorImpl();
-    clientSocketCreator = new ClientSocketCreatorImpl();
-    advancedSocketCreator = new AdvancedSocketCreatorImpl();
+    serverSocketCreator = new ServerSocketCreatorImpl(this);
+    clientSocketCreator = new ClientSocketCreatorImpl(this);
+    advancedSocketCreator = new AdvancedSocketCreatorImpl(this);
   }
 
 
@@ -66,7 +45,7 @@ public class TcpSocketCreatorImpl implements TcpSocketCreator {
     return false;
   }
 
-  private Socket connect(HostAndPort addr, int timeout,
+  Socket connect(HostAndPort addr, int timeout,
       ConnectionWatcher optionalWatcher, boolean clientSide, int socketBufferSize)
       throws IOException {
     return forAdvancedUse().connect(addr, timeout, optionalWatcher, clientSide, socketBufferSize,
@@ -87,236 +66,6 @@ public class TcpSocketCreatorImpl implements TcpSocketCreator {
   @Override
   public AdvancedSocketCreator forAdvancedUse() {
     return advancedSocketCreator;
-  }
-
-  protected class ServerSocketCreatorImpl implements ServerSocketCreator {
-
-    public boolean useSSL() {
-      return TcpSocketCreatorImpl.this.useSSL();
-    }
-
-    public void handshakeIfSocketIsSSL(Socket socket, int timeout) throws IOException {
-      if (useSSL()) {
-        throw new IllegalStateException("Handshake on SSL connections is not supported");
-      }
-    }
-
-    @Override
-    public final ServerSocket createServerSocket(int nport, int backlog) throws IOException {
-      return createServerSocket(nport, backlog, null, -1, useSSL());
-    }
-
-    @Override
-    public final ServerSocket createServerSocket(int nport, int backlog, InetAddress bindAddr)
-        throws IOException {
-      return createServerSocket(nport, backlog, bindAddr, -1, useSSL());
-    }
-
-    /**
-     * Overridable method for creating a server socket. Override this if you are implementing
-     * SSL communications or otherwise need to customize server socket creation.
-     */
-    protected ServerSocket createServerSocket(int nport, int backlog, InetAddress bindAddr,
-        int socketBufferSize, boolean sslConnection) throws IOException {
-      if (sslConnection) {
-        throw new UnsupportedOperationException();
-      }
-      ServerSocket result = new ServerSocket();
-      result.setReuseAddress(true);
-      if (socketBufferSize != -1) {
-        result.setReceiveBufferSize(socketBufferSize);
-      }
-      try {
-        result.bind(new InetSocketAddress(bindAddr, nport), backlog);
-      } catch (BindException e) {
-        BindException throwMe =
-            new BindException(String.format("Failed to create server socket on %s[%s]",
-                bindAddr == null ? InetAddress.getLocalHost().getHostAddress() : bindAddr,
-                nport));
-        throwMe.initCause(e);
-        throw throwMe;
-      }
-      return result;
-    }
-
-    /**
-     * Return a client socket. This method is used by peers.
-     */
-    public Socket connectForServer(HostAndPort addr) throws IOException {
-      return connect(addr, 0, null, false, -1);
-    }
-
-    @Override
-    public final Socket connectForServer(HostAndPort addr, int timeout,
-        ConnectionWatcher optionalWatcher)
-        throws IOException {
-      return forAdvancedUse().connect(addr, timeout, optionalWatcher, false, -1, useSSL());
-    }
-
-  }
-
-  protected class ClientSocketCreatorImpl implements ClientSocketCreator {
-    @Override
-    public boolean useSSL() {
-      return TcpSocketCreatorImpl.this.useSSL();
-    }
-
-    @Override
-    public void handshakeIfSocketIsSSL(Socket socket, int timeout) throws IOException {
-      if (useSSL()) {
-        throw new IllegalStateException("Handshake on SSL connections is not supported");
-      }
-    }
-
-    /**
-     * Return a client socket. This method is used by client/server clients.
-     */
-    public Socket connectForClient(HostAndPort addr, int timeout) throws IOException {
-      return connect(addr, timeout, null, true, -1);
-    }
-
-    /**
-     * Return a client socket. This method is used by client/server clients.
-     */
-    public Socket connectForClient(HostAndPort addr, int timeout, int socketBufferSize)
-        throws IOException {
-      return connect(addr, timeout, null, true, socketBufferSize);
-    }
-
-  }
-
-  protected class AdvancedSocketCreatorImpl implements AdvancedSocketCreator {
-    @Override
-    public boolean useSSL() {
-      return TcpSocketCreatorImpl.this.useSSL();
-    }
-
-    @Override
-    public void handshakeIfSocketIsSSL(Socket socket, int timeout) throws IOException {
-      if (useSSL()) {
-        throw new IllegalStateException("Handshake on SSL connections is not supported");
-      }
-    }
-
-    @Override
-    public Socket connect(HostAndPort addr, int timeout,
-        ConnectionWatcher optionalWatcher, boolean allowClientSocketFactory,
-        int socketBufferSize, boolean useSSL) throws IOException {
-      if (useSSL) {
-        throw new IllegalArgumentException();
-      }
-      Socket socket = null;
-      if (allowClientSocketFactory) {
-        socket = createCustomClientSocket(addr);
-      }
-      if (socket == null) {
-        socket = new Socket();
-
-        // Optionally enable SO_KEEPALIVE in the OS network protocol.
-        socket.setKeepAlive(ENABLE_TCP_KEEP_ALIVE);
-
-        if (socketBufferSize != -1) {
-          socket.setReceiveBufferSize(socketBufferSize);
-        }
-        if (optionalWatcher != null) {
-          optionalWatcher.beforeConnect(socket);
-        }
-        InetSocketAddress inetSocketAddress = addr.getSocketInetAddress();
-        try {
-          InetAddress serverAddress = inetSocketAddress.getAddress();
-          if (serverAddress == null) {
-            serverAddress = InetAddress.getByName(inetSocketAddress.getHostName());
-          }
-          socket.connect(
-              new InetSocketAddress(serverAddress, inetSocketAddress.getPort()),
-              Math.max(timeout, 0));
-        } finally {
-          if (optionalWatcher != null) {
-            optionalWatcher.afterConnect(socket);
-          }
-        }
-      }
-      return socket;
-    }
-
-    @Override
-    public final ServerSocket createServerSocketUsingPortRange(InetAddress ba, int backlog,
-        boolean isBindAddress, boolean useNIO,
-        int tcpBufferSize, int[] tcpPortRange,
-        boolean sslConnection) throws IOException {
-      try {
-        // Get a random port from range.
-        int startingPort = tcpPortRange[0]
-            + ThreadLocalRandom.current().nextInt(tcpPortRange[1] - tcpPortRange[0] + 1);
-        int localPort = startingPort;
-        int portLimit = tcpPortRange[1];
-
-        while (true) {
-          if (localPort > portLimit) {
-            if (startingPort != 0) {
-              localPort = tcpPortRange[0];
-              portLimit = startingPort - 1;
-              startingPort = 0;
-            } else {
-              throw noFreePortException(
-                  String.format("Unable to find a free port in the membership-port-range: [%d,%d]",
-                      tcpPortRange[0], tcpPortRange[1]));
-            }
-          }
-          ServerSocket socket = null;
-          try {
-            if (useNIO) {
-              ServerSocketChannel channel = ServerSocketChannel.open();
-              socket = channel.socket();
-
-              InetSocketAddress address =
-                  new InetSocketAddress(isBindAddress ? ba : null, localPort);
-              socket.bind(address, backlog);
-            } else {
-              socket = ((ServerSocketCreatorImpl) forServer()).createServerSocket(localPort,
-                  backlog, isBindAddress ? ba : null,
-                  tcpBufferSize, sslConnection);
-            }
-            return socket;
-          } catch (java.net.SocketException ex) {
-            if (socket != null && !socket.isClosed()) {
-              socket.close();
-            }
-            localPort++;
-          }
-        }
-      } catch (IOException e) {
-        throw problemCreatingSocketInPortRangeException(
-            "unable to create a socket in the membership-port range", e);
-      }
-    }
-
-    /**
-     * Overridable method for creating an exception during search of port-range
-     */
-    protected RuntimeException problemCreatingSocketInPortRangeException(String s, IOException e) {
-      return new RuntimeException(s, e);
-    }
-
-    /**
-     * Overridable method for creating an exception during search of port-range
-     */
-    protected RuntimeException noFreePortException(String reason) {
-      return new RuntimeException(reason);
-    }
-
-    /**
-     * reimplement this method to use a custom socket factory to create and configure a new
-     * client-side socket
-     *
-     * @return the socket, or null if no custom client socket factory is available
-     */
-    protected Socket createCustomClientSocket(HostAndPort addr) throws IOException {
-      throw new UnsupportedOperationException(
-          "custom client socket factory is not supported by this socket creator");
-    }
-
-
   }
 
 }
